requirements.txt
langchain
openai
python-dotenv
requests 
beautifulsoup4 
langchain 
langchain-community 
langchain-core 
python-dotenv 
langgraph
langchain-openai